[NET-952][NET-955] Refactor exportd frontend module
