+++ app/src/app/exportd/exportd-job-settings-builder/exportd-job-destinations-step/exportd-job-destinations-step.component.scss
@@ -1,4 +1,4 @@
-@import "~src/scss/variables";
+@import "../../../../scss/variables";
 
 .card-header {
   button {
