+++ app/src/app/exportd/exportd-job-settings-builder/exportd-job-variables-step/exportd-job-variables-step.component.scss
