+++ app/src/app/layout/structure/navigation/navigation.component.html
@@ -65,7 +65,7 @@
                   <div class="row no-gutters">
                     <div class="col">
                       <a class="dropdown-icon-item" *permissionLink="['base.exportd.job.view']"
-                         routerLink="/settings/exportdjob">
+                         [routerLink]="['/exportd']">
                         <i class="fas fa-tachometer-alt"></i> <span>Exportd</span>
                       </a>
                     </div>
