+++ app/src/app/main/main-routing.module.ts
@@ -92,6 +92,8 @@ const routes: Routes = [
     data: {
       breadcrumb: 'DocAPI'
     },
+    canActivate: [AuthGuard],
+    canActivateChild: [AuthGuard],
     loadChildren: () => import('../docapi/docapi.module').then(m => m.DocapiModule)
   },
   {
@@ -120,6 +122,15 @@ const routes: Routes = [
     canActivate: [AuthGuard],
     canActivateChild: [AuthGuard],
     loadChildren: () => import('../filemanager/filemanager.module').then(m => m.FilemanagerModule)
+  },
+  {
+    path: 'exportd',
+    data: {
+      breadcrumb: 'Exportd'
+    },
+    canActivate: [AuthGuard],
+    canActivateChild: [AuthGuard],
+    loadChildren: () => import('../exportd/exportd.module').then(m => m.ExportdModule)
   },
   {
     path: 'debug',
