+++ app/src/app/settings/log-settings/log-settings.component.html
@@ -1,7 +1,7 @@
 
 <div class="card-deck">
   <div class="card">
-    <a routerLink="exportdjobs">
+    <a routerLink="exportd">
       <div class="dash-card primary">
         <i class="fas fa-tachometer-alt"></i>
         <span class="count-numbers">Exportd</span>
